AutoMigrate Database

+++ entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID       uint   `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID       uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name     string `gorm:"type: varchar(255)" json:"name"`
 	Email    string `gorm:"uniqueIndex; type:varchar(255)" json:"email"`
 	Password string `gorm:"->:false;<-;not null" json:"-"`
